MB-58317 Improve bucket updater logging
This improves the logging from the bucket updater such that a normal
exit when a bucket is removed is not reported in the query.log as an
error level message.

It also removes a redundant sequences clean-up call and stops repeated
calls to clean-up sequences & scope functions when the bucket is not
found, thereby reducing the number of "not found" messages logged (plus
saves the wasted processing trying to repeatedly find the bucket).

Change-Id: Ia9a2d5441ec1c5797c3be518bab4dd953bac38dd
Reviewed-on: https://review.couchbase.org/c/query/+/195708
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>