MB-58356 Synchronise access to scopes list
The clean-up for sequences invoked when a keyspace is loaded exposed
the fact that whilst we lock the keyspace when making changes to the
scopes list, we don't lock it similarly when using the scopes list
exposing us to panics when the list changes under a reader.

This both delays the clean-up until the keyspace entry has been fully
loaded and included in the namespace's list, and adds synchronisation
to scopes list read access.

Change-Id: I01efb7e2368adfc86d0f282e22d7b5ee7f34e6d2
Reviewed-on: https://review.couchbase.org/c/query/+/195864
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>