MB-58317 Revise fix
The previous fix would inadvertently result in other buckets being
unloaded/deleted on pool refresh.

This revision addresses this and revises the messages produced to more
easily tie them to an instance of a bucket by including an abbreviated
identifer in the message prefix.  The abbreviated identifier is the
first 4 characters of the bucket name (underscore padded) and the first
and last 4 digits of the bucket's UUID.  (Abbreviated to try keep
messages from growining illegibly long.)

Change-Id: I539994e691ce999cc55075e7a65e0a633215755d
Reviewed-on: https://review.couchbase.org/c/query/+/195980
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>