MB-57717 Drop N1QL_SYSTEM_BUCKET at end of migration
Add logic to drop N1QL_SYSTEM_BUCKET at the end of migration.
Get a document count and if it is not the expect 0, gather
the document keys and print in the log. Then drop
N1QL_SYSTEM_BUCKET. The drop happens regardless of whether
there are any unexpected remaining documents in
N1QL_CBO_STATS, since we don't have a manual way of doing
moving the documents anyway, and migration is marked as
"complete" so no automatic way of moving either.

Change-Id: Icf834a0fe24d6b1524a450d0b53ebb778540d39e
Reviewed-on: https://review.couchbase.org/c/query/+/195986
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>