MB-57778 Set num CPUs only on start
Previously in MB 56068 we permitted adjusting CPUs dynamcially through
settings.  Since this now passes the settings value on the command line
we can restrict changes to start-up (less risk).  This just replaces the
setting processor with a debug level only message so there is no issue
with the setting being passed dynamically - we just ingore it.

Change-Id: Id123b4c3e34d520c5206e97b7e8ca6ac0f13c77f
Reviewed-on: https://review.couchbase.org/c/query/+/196033
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>