MB-58435 Quota accounting improvements
1. Make sure that the fetched document's size is recalculated to reflect
the data fetched that is added to the empty original.

2. In order.go, handle growth of the document in makeMinmal.  The document
could grow (despite the function's attempt to shrink it) if the ORDER BY
terms generate large values, e.g. ORDER BY array_repeat(type, 10000)

3. Reset the original value only after recalculating the size in
makeMinimal.  This is because we don't actually free the original in
case there are dangling un-tracked field references.  But we don't want
to reduce the quota use if the original is hanging around -- better to
be over than under with the quota accounting.

4. Include the size of the original value in the size of an annotated
value. (see #3)

5. Include the attachments in an annotated value's size even if shared -
better to double count than fail to count.  Also, if we don't include
them always the values generated in #2 may not be included in the quota
use calculation.  (Typically attachments are flagged as shared with the
original value in InitialProject - but then neither account for these
generated order term values.)

6. Have ScopeValues include the parent's size in their size.  They may
be the last reference to the parent and therefore that memory is still
in use whilst the ScopeValue exists and must be accounted for.  Again,
better to count twice than not at all.

Change-Id: I230235f1707ed5702d59923a37471e6c8fc28d34
Reviewed-on: https://review.couchbase.org/c/query/+/196351
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>