MB-57781 Modify fix
The previous fix added save/restore of this.node in VisitSubselect,
and set this.node to the Subselect. Although this has been the way
we set this.node since the original implementation of extended
attributes (xattrs), there is an issue with this in that the
Subselect does not include information on ORDER BY, i.e., if a
query references xattrs in the ORDER BY clause, we won't be able
to properly identify the xattrs there.

The fix is to remove the logic for save/restore/set of this.node
in Subselect. We already do that in Select (which does include
ORDER BY info). In case of set operations (UNION/INTERSECT/EXCEPT)
it's ok to save/restore/set for each arm of the set operations
(which is done in build_set_op.go, no need to do it in Subselect),,
since in that case the ORDER BY cannot reference the underlying
keyspaces in individual arms anymore (can only reference
projection aliases).

Change-Id: If89bbd85a12e95c674f035c068735b640e6201e7
Reviewed-on: https://review.couchbase.org/c/query/+/196332
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>