MB-58465 early projection
Implement early projection at the Fetch after the document
has been fetched. The early projection is a list of
top-level field names gathered from all the expressions
in the statement, if neither a star nor an entire keyspace
is referenced in the expressions. After the early
projection list is obtained, make sure the list can
"cover" the query for this keyspace.
At runtime if early projection is done, start with an
empty document and only get the value for the fields in
early projection and set the value in the new document.

Change-Id: Iaf380e0729fee649b2377024bbf5dad210c16376
Reviewed-on: https://review.couchbase.org/c/query/+/196380
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>