MB-58106: Fix covering bug in queries with CTEs
A query with an upstream CTE - is mistakenly not marked as
covered if the CTE references the same keyspace alias
as the outer query. Even if there is a covering index.

Upstream CTEs should not be included in Covering Index Check
or in the Covering Transformation.

* Only include subqueries that are correlated with the
current keyspace during the covering index check

* In cover transformation for a 'Select' that is being planned
  - only include With expressions that are evaluated Downstream

  An indicator in algebra.Select is used to indicate if With
  expressions should be included in cover transformation.

Example:

Outer query is not covered - even if there is a covering index.

CREATE INDEX ON kids(id);

WITH cte1 AS ( SELECT name FROM kids WHERE id = 1 )
SELECT id, cte1 FROM kids
WHERE id = 2
ORDER BY id;

Change-Id: Idf6479522f48f6836e0bdb2d8dc2dd239a1de492
Reviewed-on: https://review.couchbase.org/c/query/+/196452
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>