MB-58150 Further limit cases we use OR clause filters for sarging index keys
The issue is a continuation of using individual OR clause to
sarg index keys. For OR-clauses we used to never use individual
OR-clause filters for sarging index keys, we instead use the
entire DNF-transformed OR clause. As part of the fix for
MB-51817 we allow using of individual OR-clause for sarging
index keys, for the benefit of marking individual OR-clause
as exact if we do so, which benefits gathering of post-index
filters as part of that feature.

There have been several issues that resulted from this, e.g.,
MB-56039, the fix for which we introduced a new function
orSargUseFilters() to limit the cases where the situations
where we should use individual OR-clause filters for sarging.

The issue with this bug is that, if we have an OR-clause that
can be used to sarg multiple index keys, e.g.:
   index on (c1, c2)
   (c1 = 1 AND c2 = 2) OR (c1 = 3 AND c2 = 4)
then if we use the individual OR clause to sarg index keys,
we'll end up with 2 spans for key c1 and 2 spans for key c2,
and then combine them into 4 spans for both index keys,
which is cross multiply of the 2 index keys. This is
incorrect since the original OR-clause only have 2 spans for
the 2 index keys.

This problem does not exist if we use the original approach
of using the entire OR clause to sarg all index keys.
The fix is to add more restrictions in determining whether
to use individual OR clause filters for sarging index keys.
If we detect that multiple sargable index keys exist from
the OR clause, fall back to the original mechanism instead of
trying to use individual OR-clause for sarging index key.

Change-Id: I72aa25e8647da5f9e8ec31c3d8b2dd1f57aa39e4
Reviewed-on: https://review.couchbase.org/c/query/+/195012
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/196748
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker