MB-58585 Sequential scan: better share resources
Limits the number of keys requested for individual v-bucket scans when
the sequential scan is ordered.  This is to reduce overall load when
most will require fewer than ~100k keys in total.  The number fetched per
v-bucket is 102400 / #v-buckets.  It is #v-buckets dependent so it works
for a variety of environments.

Change-Id: Iefa8415591f8edd41637156c48e9c39e2740b8d0
Reviewed-on: https://review.couchbase.org/c/query/+/196896
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>