MB-58435 Further updates to quota accounting...
... and memory use tweaks.

These changes are intended to nudge the request quota figure closer to
actual memory usage.  Of course this figure can't account for garbage
lag so RSS will never track it accurately, but these aim to for it to be
closer.  The most likely side-effect is suddenly seeing a marked
increase in the usedMemory figure and/or the likelihood of hitting a
"quota exceeded" error when previously one wouldn't have been raised.

In addition it changes the free space used to calculate when to spill
(order & group by) to include buffers.  This is so we better use memory
and don't rush to spill when unnecessary.

It also introduces an attempt to return memory to the OS should the GC
not have recently run and the OS free memory (including buffers) is less
than 25%.  This will typically not kick in while the server is busy but
may mean idle servers return memory to the OS a bit sooner.

Change-Id: Ie1877a2a6c6aa1b30e1e2c650e8afd9fb0917058
Reviewed-on: https://review.couchbase.org/c/query/+/196651
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>