MB-58607 Proper copy of Timeseries expression
Timeseries is a built-in function (expression), and when the
Timeseries expression is being copied, we did not set this.expr
properly on the newly copied function expression, which results
in nil pointer exception later on.

The fix is to properly set this.expr of the copied expression.
Also optimize the costing of UNNEST to check whether the unnest
expression is indexable, if not, there is no need to probe
histograms.

Change-Id: I74b659ae858ff5bb902c8e5126eae722bd8b85c4
Reviewed-on: https://review.couchbase.org/c/query/+/197051
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/197053
Well-Formed: Restriction Checker