MB-58479 [7.1.5.MP4]: Fix fatal errors...
on concurrent Inline function execution

Since the same AST object is shared across Inline function
executions:

1. When an inline function that contains subqueries in its function body
is concurrently executed - there can be fatal concurrent operations
and race conditions.

2. In some cases, subsequent executions of the function can give
   wrong results/ errors due to previous transformations on the same
   shared AST

To fix this:

When an inline function that contains subqueries
is executed:

* Generate a new AST object by re-parsing the inline body's expression
  and use this new object for further planning and processing.

* For every query, maintain a map in the execution context:
    Key: full path of the inline UDF
    Value: Actual AST object of UDF
           Local AST object to use for current execution

* Since the same query can execute a function multiple times:
  Example: SELECT inline1() FROM ARRAY_RANGE(0,10) a;
  When generating the AST expression for function execution
  check if the function's AST body has changed from the entry in the map
  in the interim. And re-create the entry if there are changes.

MB-58479 [7.1.5.MP4]: Mark subqueries as inside function

For inline functions the previous fix will reparse the function
expression at execution time if it contains subqueries. The
issue is that for the original expression, we also go through
a formalization process that could potentially mark a subquery
as being inside a function, such that at runtime it'll not try
to use cached subquery result. When we reparse and generate a
new expression we are missing this step of marking subqueries
and thus may incorrectly try to use cached subquery result at
execution time.

The fix is to add an extra step of gathering all subqueries and
mark each one as "inside function" when the function has
arguments.

MB-58479 [7.1.5.MP4]: Do not share execution tree for subqueries inside inline functions

Currently plans and executions trees for subqueries are cached in execution
context and can be shared. This works ok for regular subqueries, since they
are "unique" in a statement. For a subquery inside an inline functions,
this could potentially be problematic since the same inline function can
potentially be called from multiple places in the statement, in which case
we should not be sharing execution tree for such subqueries.

We have an existing indicator in algebra.Select when a statement is inside
inline function, however this is currently only used if the inline function
has parameters. To fix this properly, we add a separate indicator in
algebra.Select that indicates whether this select statement is inside an
inline function. This is in addition to the existing indicator for
hasVariables. Modify the formalizer code and the code for processing
subqueries inside inline functions (when the function expression is reparsed)
such that we have proper markers on each subquery inside an inline function.
Then when a subquery is evaluated at runtime, disable sharing of execution
tree for such subqueries. (query plan for subqueries can still be shared).

MB-58479 [7.1.5.MP4]: Use the same mutex for fields shared among multiple contexts

There are a few fields in Context that are shared when the query context
is copied (see NewQueryContext() function). These currently include:
subExecTrees, udfPlans, udfStmtExecTrees, inlineUdfExprs. The access for
these fields are currently protected by the embedded mutex in the context,
however this is problematic since these fields are shared among multiple
contexts and thus using embedded mutext in a context is not sufficient.

To fix this, we introduce a new queryMutex that is a pointer, and when
we copy over the above-mentioned fields we also copy over the queryMutex,
such that the same mutex is used to protect the shared fields.

Change-Id: Ia10a43a43035837d58d8e0e5752f253740fa5eb3
Reviewed-on: https://review.couchbase.org/c/query/+/197307
Well-Formed: Restriction Checker
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>