MB-58465 Control use of sub-doc API with param
Use the request-level "n1ql_feat_ctrl" setting to control use of the
sub-document API for fetch operations.  Performance suffers with its
use but this may be outweighed by the reduction data transferred between
Query and the KV in some circumstances.  When the sub-document API is
not in use Query may still perform early projection.

Change-Id: I18a9f37be52fc9ad520e3d9874ab61db9b4e0866
Reviewed-on: https://review.couchbase.org/c/query/+/197297
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>