MB-58479 Mark subqueries as inside function
For inline functions the previous fix will reparse the function
expression at execution time if it contains subqueries. The
issue is that for the original expression, we also go through
a formalization process that could potentially mark a subquery
as being inside a function, such that at runtime it'll not try
to use cached subquery result. When we reparse and generate a
new expression we are missing this step of marking subqueries
and thus may incorrectly try to use cached subquery result at
execution time.

The fix is to add an extra step of gathering all subqueries and
mark each one as "inside function" when the function has
arguments.

Change-Id: I2812ed516f838b74b7875001ef047b4c5fc95971
Reviewed-on: https://review.couchbase.org/c/query/+/197350
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>