MB-58713 Reduce the number of servicers...
...when available memory is low.

Limitations:
Most notably this can only react to high-but-not-fatal memory use.  As a
result the percieved service becomes inconsistent where sometimes a
reqest will fail but other times it will just be delayed.  The duration
of the delay for a specific request can vary when trying to repeat a
scenario.

Change-Id: Ie2db3b12b83af8363d858ae993c266e9c1adf24e
Reviewed-on: https://review.couchbase.org/c/query/+/195809
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>