MB-58719 Avoid setting cbKeyspace to nil in case of _NEEDS_REFRESH
If a keyspace needs to be reloaded due to _NEEDS_REFRESH flag
(or _DELETED flag), current code set entry.cbKeyspace to nil, such
that it can be reloaded. The issue with this is that first, the
setting of cbKeyspace to nil is not under entry.Lock, thus it'll
not prevent nil pointer access, thus in between the time this is
set to nil to the time when a new keyspace is successfully loaded,
any existing reference to the original keyspaceEntry could be
accessing the nil pointer. Our code seems to assume that if a
keyspaceEntry exists and being used, the cbKeyspsace pointer is
valid, we access the pointer in many places without checking for
nil pointer. Thus setting cbKeyspace to nil seems to be a bad
idea in general (we have a comment later on in the code that
says this is the only place where we set cbKeyspace, which is
obviously not true if we set it to nil early one).

Modify the logic for (re)loading cbKeyspace such that if
_NEEDS_REFRESH or _DELETED flag is set, we remember the old
cbKeyspace, and determine later on whether the cbKeyspace pointer
on the keyspaceEntry is already reloaded by someone else, or was
it still the original pointer.

Change-Id: I172559ae32b0b397a6995b215ea7bb2cbf5e8d84
Reviewed-on: https://review.couchbase.org/c/query/+/197374
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
(cherry picked from commit 56da3f769a37c9a1527f193a684aff450a0f9e82)
Reviewed-on: https://review.couchbase.org/c/query/+/196718
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker