MB-58479 Do not share execution tree for subqueries inside inline functions
Currently plans and executions trees for subqueries are cached in execution
context and can be shared. This works ok for regular subqueries, since they
are "unique" in a statement. For a subquery inside an inline functions,
this could potentially be problematic since the same inline function can
potentially be called from multiple places in the statement, in which case
we should not be sharing execution tree for such subqueries.

We have an existing indicator in algebra.Select when a statement is inside
inline function, however this is currently only used if the inline function
has parameters. To fix this properly, we add a separate indicator in
algebra.Select that indicates whether this select statement is inside an
inline function. This is in addition to the existing indicator for
hasVariables. Modify the formalizer code and the code for processing
subqueries inside inline functions (when the function expression is reparsed)
such that we have proper markers on each subquery inside an inline function.
Then when a subquery is evaluated at runtime, disable sharing of execution
tree for such subqueries. (query plan for subqueries can still be shared).

Change-Id: I5bd951b4cdd84c287b26cd9bbb922feda15d1f0e
Reviewed-on: https://review.couchbase.org/c/query/+/197390
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>