MB-58479 Use the same mutex for fields shared among multiple contexts
There are a few fields in Context that are shared when the query context
is copied (see NewQueryContext() function). These currently include:
subExecTrees, udfPlans, udfStmtExecTrees, inlineUdfExprs. The access for
these fields are currently protected by the embedded mutex in the context,
however this is problematic since these fields are shared among multiple
contexts and thus using embedded mutext in a context is not sufficient.

To fix this, we introduce a new queryMutex that is a pointer, and when
we copy over the above-mentioned fields we also copy over the queryMutex,
such that the same mutex is used to protect the shared fields.

Change-Id: I6e6e8eea54db1c7021dcf63a7f1c64e4e1b502d8
Reviewed-on: https://review.couchbase.org/c/query/+/197391
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>