MB-58648: Revise handling of profile info for subqueries in Inline UDFs
For Inline UDFs - since the same UDF can be called multiple times
in the same query - and since we do not allow sharing for these
subqueries - multiple invocations can cause leaks in the subquery
execution tree
 cache in the context

For subqueries inside inline UDFs -
since we do not allow sharing of the execution tree - multiple
invocations of the same UDF in a query can cause memory leaks in
the subquery execution tree cache in the execution context.

To fix this:
1. For a subquery that is inside an Inline UDF - store only 1 execution
tree in the cache for profiling purposes.

2. Since a new tree is created for every execution of such a subquery -
- accumulate its timing info in the existing tree and then clean up
the new created tree.

Change-Id: I0e6897aab5d4b8a1af0f887d00d12a21bd287141
Reviewed-on: https://review.couchbase.org/c/query/+/197426
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>