MB-58720 Add panic recovery to GenCache
The idea here is that should we have a panic in the nonBlocking function
(e.g. whilst traversing actives) we release the held lock.  A panic
elsewhere does nothing.  (All panics are passed on to the caller.)

This is just for additional safety so we don't leave a server seemingly
up but dead to requests.

Change-Id: I52a81aca01641c705a96d37e4e68f886878544d4
Reviewed-on: https://review.couchbase.org/c/query/+/197525
Tested-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>