MB-58779 [BP MB 58720] Add panic recovery to GenCache
The idea here is that should we have a panic in the nonBlocking function
(e.g. whilst traversing actives) we release the held lock.  A panic
elsewhere does nothing.  (All panics are passed on to the caller.)

This is just for additional safety so we don't leave a server seemingly
up but dead to requests.

Change-Id: I5b96ec5b37bb5d1fc142763b00349080c98c2af7
Reviewed-on: https://review.couchbase.org/c/query/+/197568
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker