MB-58106, MB-58246: [BP to 7.2.3]
MB-58106: Fix covering bug in queries with CTEs

A query with an upstream CTE - is mistakenly not marked as
covered if the CTE references the same keyspace alias
as the outer query. Even if there is a covering index.

Upstream CTEs should not be included in Covering Index Check
or in the Covering Transformation.

* Only include subqueries that are correlated with the
current keyspace during the covering index check

* In cover transformation for a 'Select' that is being planned
  - only include With expressions that are evaluated Downstream

  An indicator in algebra.Select is used to indicate if With
  expressions should be included in cover transformation.

Example:

Outer query is not covered - even if there is a covering index.

CREATE INDEX ON kids(id);

WITH cte1 AS ( SELECT name FROM kids WHERE id = 1 )
SELECT id, cte1 FROM kids
WHERE id = 2
ORDER BY id;

----

MB-58246: Fix query not being marked as covered...

if it contains a subquery that has a Self /
un-prefixed *  reference

Example:
CREATE INDEX ix1 ON kids(id);

EXPLAIN SELECT id, ( SELECT * FROM people) FROM kids WHERE id IS NOT MISSING;

The outer query is not covered even though there is a covering index.

Change-Id: I6c44bc3702adbe1894dd5a4f337b9f2a60305e83
Reviewed-on: https://review.couchbase.org/c/query/+/197605
Well-Formed: Restriction Checker
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>