MB-58648 : [ BP MB 56432 to 7.1.x] Load privileges at function creation
This leads to RBAC not being honoured.

* Fix RBAC not being honoured in some scenarios.

Because we now try not to re-load the function from storage at
each execution - in some scenarios the privileges were never loaded.
This adds a fix for it.

* Fixes misleading Authorization failure error.

Check for the Execute Function privileges before the privileges required
for queries inside the function body.

Change-Id: I4ce16f115f95d18a5defffd16d6e1a00db7cd638
Reviewed-on: https://review.couchbase.org/c/query/+/197739
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>