MB-57717 Rework migration - part 1 (single node)
The migration infrastructure has been redesigned due to a flaw
in previous assumption. We cannot effectively use metakv as a
mechansim for synchronization, since metakv is eventual
consistent only. The previous design for a single query node
to perform migration cannot be reliably achieved due to this
design flaw.

Instead of designating one query node for performing migration,
the new design will allow all query nodes to perform migration
and the point of "synchronization" will be the document-level
atomicity of each item. For UDF migration, each item is an
entry in previous metakv path; for CBO stats migration, each
item is a document in N1QL_CBO_STATS (under N1QL_SYSTEM_BUCKET).

The UDF migration no longer waits for all buckets to have
system collection then do it in one swoop. It now migrates on
a bucket-by-bucket basis. This means that there will be multiple
scans of the metakv location which might be slow, but we don't
really have an alternative in the new scheme of parallel
migrations.

The CBO stats migration also can no longer migrate on a per
keyspace basis (i.e. gather all relevant documents for a
keyspace, load up a dictionary cache entry for the keyspace,
then write all documents to new location). Instead it need to
migrate on a per document basis.

The first part of the change is to get the basic migration to
work on a single node setup, i.e. no multiple nodes performing
migration at the same time. This is just to get the basic
infrastructure in place for the new migration scheme.

Change-Id: Iab2b1bc2f39bb5a4a3eb8dd5d0f3c54356c439dc
Reviewed-on: https://review.couchbase.org/c/query/+/197988
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>