MB-57717 Rework migration - part 3 (error handling)
Since migration happens in parallel (all query nodes) we need
to be a bit smarter about error handling, i.e. distinguishing
which errors could be expected and which errors are not.

It turns out for Fetch and Delete, we do not return an error
if the key does not exist (the datastore layer explicitly
ignores key not found error), thus we don't really have any
"expected" error for these two operators. For Insert, the only
"expected" error is duplicate key (which gets translated to
duplicate function in UDF).

We now detect and return errors during migration process
(except for "expected" errors). This allows us to halt
migration in case of unexpected errors.

We still need to add mechanism for retry migration in case
of errors. That's coming up next.

Change-Id: Iad67a26906513e52f330aeec8e14162913ac2d39
Reviewed-on: https://review.couchbase.org/c/query/+/198315
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>