MB-58435 Revise changes
The original tweaks in MB-58435 removed the independent id field but
failed to ensure it was preserved when meta was reset.  They also
exposed the fact that shared annotations were being cleared from under
the shared user.

This both preserves the id and ensures that both parties in shared use
are flagged such that they don't clear the shared attributes from under
the other party.

Change-Id: I9e5efebbabd33aed45a06be388e3df20f8ea0328
Reviewed-on: https://review.couchbase.org/c/query/+/198353
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>