MB-57717 Block access early during migration
The new scheme for migration is to have all query nodes attempt to
do migration when possible. One side effect of this approach is
that when a query node starts the migration process, it has no
information on what another query node might have done, i.e.
whether parts of the old information has been migrated to the new
location. As a result, the query node cannot reliably preload
the relevant caches since the old location may only contain
partial information. This same limitation also applies to any
potential query run at this time that needs access to the required
information (e.g. function definition, or CBO stats). Therefore,
we need to block such access as soon as the migration process
starts, by changing the migration state to _MIGRATING.

In case of bucket-by-bucket migration through the registered
migrator, this is already done, i.e. the first bucket that has the
new capability will change the migration state. In case of
migrateAll(), move the changing of migration state to be the first
thing it does.

Change-Id: I4afe3c5e887fb24d8945cb542a9348a24b2b9b8c
Reviewed-on: https://review.couchbase.org/c/query/+/198398
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>