MB-57717 Improve index scan efficiency for system collection
Currently when we scan _query collection under _system scope, for either
UDF definition or CBO stats, we use a full index scan on the primary
index. Since the document keyss under there all have a relevant prefix
(cbo:: for CBO stats documents and udf:: for UDF documents), we could
at least use the prefix as index spans such that we don't need to scan
the entire primary index.

The CBO stats code already have a mechanism for generating index spans,
for cases where the documents are under N1QL_SYSTEM_BUCKET. We simply
use the same mechanism and provide the prefix (cbo::) as document key
to the mechanism. For UDF currently it uses primary scan directly,
replace that with the same mechanism used in CBO stats code.

Also modify the processing function for scanning the index (UDF) to
return errors.Error instead of error, to be consistent with similar
function for CBO stats.

Change-Id: Idd8afb6aa50cb23c2a933c5082d25d47d3cd17ef
Reviewed-on: https://review.couchbase.org/c/query/+/198410
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>