Fix merge tests in unit test
When the merge test case was originally written we frequently
use the same keyspace as source and target, and later on we
rewrote some of the test cases to use a subquery as the source
instead. However, in many of these test cases we still refer
to the original meta(s).id in the INSERT clause of the MERGE
statement, even though s is now a subquery. The tests continue
to work due to a bug (?) in the code that inherits the meta
information in the Alias operator (since we share annotations).

Fix the tests to explicitly project the metaid in the subquery,
and refer to that instead of meta(s).id in the INSERT clause.

Change-Id: Ic315c8f7d21e0da914dab4d493a99edeae01b5e0
Reviewed-on: https://review.couchbase.org/c/query/+/198505
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>