MB-57717 Add migration abort state
Although this should not happen frequently, there is nothing that
prevents an ABORT command for migration happening at the same time
that regular migration is proceeding. To handle this properly, we
added two new migration states, _ABORTING and _ABORTED, which
marks the beginning/ending of an abort command, respectively.
Add handling of the new states in regular migration process.
In addition, the marker in metakv also now includes both _MIGRATED
and _ABORTED.

Also separate node migration state from bucket migration state,
to be more explicit.

Change-Id: I666ec0284a33492b5dfe73201b956be15065c758
Reviewed-on: https://review.couchbase.org/c/query/+/198593
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>