MB-59538 Proper parent value setup for inner of nested-loop join
For the inner of a nested-loop join, currently the generated value
does not include the "parent", since the parent value includes all
terms on the outer of the nested-loop join (e.g l JOIN r, we don't
want the value for r to include as nested scope value for l). As a
side effect of this, we also lose all other kinds of parent values,
such as correlation values from a parent query, or function
arguments from inline UDF, or a WITH clause.

To fix this, for the right-hand side of a nested-loop join,
instead of always use nil as parent in the generated value from
IndexScan3, we get the parent of the "parent" value, which should
represent the original parent value.

This change also allow us to remove one of the previous restrictions
in generating a filter for either a Fetch (non-covering) or an index
scan (covering) if the keyspace is inside a correlated subquery.

Somewhat related to this, for an Alias operator, we currently only
inherit the parent value if it is the primary term; this could also
have similar issues in case we have a filter on the SubqueryTerm that
references a correlation value or a function argument. We now
unconditionally inherit the parent value for Alias operator.

Change-Id: Ie26c6dc347bcd380279f254452083abfeafdc6ae
Reviewed-on: https://review.couchbase.org/c/query/+/198721
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/198743
Well-Formed: Restriction Checker
Reviewed-on: https://review.couchbase.org/c/query/+/200624