MB-59033 Check state of primary index on system collection
There is an existing function that checks for presence of the
system collection, and for the primary index on the system
collection. Currently logic will create the primary index if
not existing, and make sure the index is online after it is
being created.

In this bug when a restore occurs the primary index on the
syste collection exists, but is in deferred mode, and attempt
to use the primary index fails.

Add logic to monitor the state of the system index if it
already exists, and if it is in deferred mode, issue the
build index command. Wait for online state of the primary
index before returns from function. Note the new logic is
not required if we create the primary index since in that
case the create function does the wait (and it won't be
in deferred mode if we created it).

Change-Id: I6664cadee25cd41a0bf43bde2b5e7037268699a4
Reviewed-on: https://review.couchbase.org/c/query/+/198747
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>