MB-57717 Make sure bucket exists before waiting for system collection
The function CheckSystemCollection() waits for the system collection
(_system._query) to show up, however, if the bucket does not exist in
the first place (e.g. in index advisor), we end up waiting for the
maximum duration unnecessarily.

Add logic to first check that the bucket exists, if not, no need to
wait.

Change-Id: I7503baebcba361a0be0d038b6c818aa39d26c711
Reviewed-on: https://review.couchbase.org/c/query/+/198752
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>