MB-59033 Scope level backup/restore...
... of functions in system storage and of sequences.

When restoring, existing functions and sequences of the same names are
replaced.  No other functions/sequences are touched - so a scope level
restore doesn't necessarily result in an exact image of what was there
at the time of the backup; new functions/sequences in the scope are not
dropped.  This differs from a full bucket restore that would result in
a new system keyspace effectively removing any new functions/sequences
not included in the backup image.

The backup format version is changed to 2 to support sequences. Backups
with a version of 1 (i.e. before this) can be restored but obviously
don't include any sequences.

Change-Id: Ie05750f70ae11efa96d7aeb88596ab3de4a2cb5a
Reviewed-on: https://review.couchbase.org/c/query/+/198789
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>