MB-57717 Allow sequential scan to be used for system collection
Currently access for the system collection (_system._query) is only
through the primary index when used for UDF and CBO stats. This
requires we check for existence and online state of the primary
index on every access. The primary index should be available for
most of the time, thus we should try to make the common access
to be more streamlined.

We now allow sequential scan to be used instead of the primary
index for accessing the system collection. We now removed the
"precheck" for primary index when we need to perform an index
scan on the system collection, and in the (hopefully) rare
cases where the primary index is not available or not online,
we return the sequential scan index instead, and in the
background issue the check system collection command which
was previously done on every access.

This allows us to undo some of the preious changes to ensure
that the primary index is available/online.

Change-Id: I324d28e48fc8e39024480e512a99feeb1724772e
Reviewed-on: https://review.couchbase.org/c/query/+/198828
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>