MB-59095 Correct pool services handling.
_POOLMAP, the master pool services variable, was storing the individual
cbPoolServices objects directly in a map.  This meant that attempts to
protect them with per-service locks were thwarted as each access
resulted in a copy of the cbPoolServices object along with its lock. A
simple fix is to rather store pointers to individual cbPoolServices
objects so all access attempts refer to the same underlying objects and
the locking is then effective.

(An alternative may be to store the lock as a pointer member, thereby
sharing just the lock but individual service objects is probably easier
to keep track of/less confusing on initial reading of the code.)

Change-Id: I788cd39f87ef42b34557157e615d853605801af1
Reviewed-on: https://review.couchbase.org/c/query/+/198955
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>