MB-57717 Revise previous fix
The cluster could be in mixed-clustered mode for an undetermined amount
of time, thus instead of a fixed number of retries for checking the
availability of system collections, keep checking and sleep every 30
seconds, and only proceed with retry of migration after at least one
bucket has started the migration process (which indicates that
migration has started). This avoids the issue of retry of migration
could get in front of the initial migration.

Change-Id: Ided55a49d1dd964a426bcab8338c4803bb356a47
Reviewed-on: https://review.couchbase.org/c/query/+/198961
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>